requirements.txt
aio-pika==9.5.7
aiormq==6.9.2
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
asyncpg==0.30.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
croniter==6.0.0
distro==1.9.0
docstring_parser==0.17.0
fast-depends==3.0.3
fastapi==0.117.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.2.0
langchain-text-splitters==0.3.11
langsmith==0.4.30
multidict==6.7.0
openai==1.109.1
orjson==3.11.3
packaging==25.0
pamqp==3.3.0
prometheus_client==0.23.1
propcache==0.4.1
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.7.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rq==2.6.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
yarl==1.22.0
zstandard==0.25.0
gitpython>=3.1.40
pathspec>=0.11.2